Archive unused useTableState shared primitive
useTableState had zero screen/view-model consumers — only a concrete barrel
re-export and its own test. Remove the hook and the re-export, leaving archive
tombstones. Its two still-used serialization types (TableSort, TableFilters)
move to their sole consumer, lib/view-state-envelope.ts. AsyncRegion is kept:
it has an active consumer in accounting-screen.close-cockpit-panels.tsx.
