+++ archive/code/README.md
@@ -33,6 +33,11 @@ surface but still explain a migration or compatibility decision.
   the legacy overview Today panel was retired with the orphaned overview screen (2026-07-02).
 - `src/Meridian.Ui/dashboard/src/screens/settings-admin-operations-console.tsx` - comment-only
   tombstone after the unrouted Settings admin operations console was retired (2026-07-02).
+- `src/Meridian.Ui/dashboard/src/components/data/use-table-state.ts` (and its test) - comment-only
+  tombstones after the unused `useTableState` shared primitive was retired (2026-07-06). It had no
+  screen or view-model consumers, only a `components/data/concrete` barrel re-export; that
+  re-export was removed. `AsyncRegion` (a sibling primitive) was kept — it has an active consumer
+  in `accounting-screen.close-cockpit-panels.tsx`.
 
 Active source remains under `src/`. Files here are reference material only and should not be
 reintroduced without moving them back into the correct active project and validating the build.
