+++ archive/code/src/Meridian.Ui/dashboard/src/components/data/use-table-state.test.ts
@@ -1,0 +1,2 @@
+// ARCHIVE TOMBSTONE
+// Retired 2026-07-06 with the unused `useTableState` primitive it covered.
